Update first task
